requirements.txt
dlib
gdown
imageio
Jinja2
networkx
ninja
opencv-python
pandas
Pillow
requests
scikit-image
scikit-learn
scipy
streamlit
torch==1.11.0
torchaudio==0.11.0
torchdiffeq==0.0.1
torchvision==0.12.0
tqdm
urllib3
wandb

